Migrate tests out of package
This reflects that the extension's tests won't be runnable from an
end-user installed state; the tests will need extra dependencies.
