+++ setup.cfg
@@ -49,7 +49,7 @@ docs =
 sphinx_example_index = data/*
 
 [tool:pytest]
-testpaths = "sphinx_example_index" "docs"
+testpaths = "tests" "docs"
 astropy_header = true
 doctest_plus = enabled
 text_file_format = rst
