+++ tests/__init__.py
@@ -1,4 +1,2 @@
 # Licensed under a 3-clause BSD style license - see LICENSE.rst
-"""
-This module contains package tests.
-"""
+"""This module contains package tests."""
